Gas optimizations

+++ src/test/utils/TicTacTokenTest.sol
@@ -18,7 +18,7 @@ contract User is ERC721Holder {
     function markSpace(
         uint256 gameId,
         uint256 i,
-        uint256 symbol
+        uint8 symbol
     ) public {
         ttt.markSpace(gameId, i, symbol);
     }
